% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{addMetaData}
\alias{addMetaData}
\alias{addMetaData.Qbone}
\title{Add in metadata associated}
\usage{
addMetaData(object, metadata, col.name = NULL)

\method{addMetaData}{Qbone}(object, metadata, col.name = NULL)
}
\arguments{
\item{object}{An object}

\item{metadata}{A vector, list, or data.frame with metadata to add}

\item{col.name}{A name for meta data if not a named list or data.frame}
}
\value{
\code{object} with metadata added
}
\description{
Add in metadata associated
}
\examples{
qbone1 <- addMetaData(object = qbone1, metadata = meta.data)
qbone1[["name"]] <- meta.data

}
