% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R, R/QboneData.R
\name{assayData}
\alias{assayData}
\alias{getQboneData}
\alias{getQboneData.Qbone}
\alias{getQboneData.QboneData}
\title{Get and Set Assay Data}
\usage{
getQboneData(object, slot, ...)

\method{getQboneData}{Qbone}(object, slot = "data", assay = NULL, ...)

\method{getQboneData}{QboneData}(object, slot = c("data"), ...)
}
\arguments{
\item{object}{An object}

\item{slot}{Specific assay data to get or set}

\item{...}{Arguments passed to other methods}

\item{assay}{Specific assay to get data from or set data for; defaults to
the \link[QboneObject:defaultAssay]{default assay}}
}
\value{
\code{getQboneData}: returns the specified assay data
}
\description{
General accessor and setter functions for \code{\link{assay}} objects.
\code{getQboneData} can be used to pull information from any of the
expression matrices (eg. \dQuote{counts}, \dQuote{data}, or
\dQuote{scale.data}). \code{SetAssayData} can be used to replace one of these
expression matrices                                              || double check
}
\concept{data-access}
