% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDots}
\alias{CheckDots}
\title{Check the Use of Dots
Adopted from Seurateobject package}
\usage{
CheckDots(..., fxns = NULL)
}
\arguments{
\item{...}{Arguments passed to a function that fall under \code{...}}

\item{fxns}{A list/vector of functions or function names}
}
\value{
Emits either an error or warning if an argument passed is unused;
invisibly returns \code{NULL}
}
\description{
Check the Use of Dots
Adopted from Seurateobject package
}
\examples{
\dontrun{
f <- function(x, ...) {
  CheckDots(...)
  return(x ^ 2)
}
f(x = 3, y = 9)
}

}
\keyword{internal}
