% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunction.R
\name{commonBasis}
\alias{commonBasis}
\title{Compute common basis}
\usage{
commonBasis(object, new.assay.name = "commonBasis", ...)
}
\arguments{
\item{object}{A Qboneobject}

\item{new.assay.name}{new assay name assigned to the commonBasis data}
}
\description{
Compute common basis
}
