% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneData.R
\name{CreateAssayObject}
\alias{CreateAssayObject}
\title{Create an QboneData object}
\usage{
CreateAssayObject(
  data,
  min.cells = 0,
  min.features = 0,
  check.matrix = FALSE,
  ...
)
}
\arguments{
\item{data}{Prenormalized data; if provided, do not pass \code{counts}}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{check.matrix}{Check counts matrix for NA, NaN, Inf, and non-integer values}

\item{...}{Arguments passed to \code{\link{as.sparse}}}
}
\value{
A \code{\link{Assay}} object
}
\description{
Create an QboneData object
}
