% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{createQboneObject}
\alias{createQboneObject}
\alias{createQboneObject.default}
\alias{createQboneObject.QboneData}
\title{Create a \code{Qbone} object}
\usage{
createQboneObject(
  data,
  project = "createQboneObject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  sampleid = 1,
  ...
)

\method{createQboneObject}{default}(
  data = data,
  project = "QboneProject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  sampleid = 1,
  ...
)

\method{createQboneObject}{QboneData}(
  data = data,
  project = "QboneProject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  sampleid = 1,
  ...
)
}
\arguments{
\item{data}{data}

\item{project}{\link{project} name for the \code{Qbone} object}

\item{assay}{Name of the initial assay}

\item{names.field}{|| double check}

\item{names.delim}{|| double check}

\item{meta.data}{Include cells where at least this many features are
detected.}

\item{sampleid}{column number of sample name in mate.data}

\item{...}{|| double check}
}
\description{
Create a \code{Qbone} object
}
\examples{
\dontrun{
n = 10000
list1 = list(rnorm(n, mean = 0, sd = 1),
             rnorm(n, mean = 0, sd = 1),
             rnorm(n, mean = 0.5, sd = 1),
             rnorm(n, mean = 0.5, sd = 1))
meta.data = data.frame(name = c("a_1", "a_2", "b_1", "b_2"),
                       group = c("a", "a", "b", "b"))
rownames(meta.data) = meta.data[,1]
qbonedata = createQboneData(list1, meta.data, sampleid = 1)
qbone1 = createQboneObject(qbonedata, meta.data = meta.data)
}

}
