% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{defaultAssay}
\alias{defaultAssay}
\alias{defaultAssay<-}
\alias{defaultAssay.Qbone}
\alias{defaultAssay<-.Qbone}
\title{Get and set the default assay}
\usage{
defaultAssay(object, ...)

defaultAssay(object, ...) <- value

\method{defaultAssay}{Qbone}(object, ...)

\method{defaultAssay}{Qbone}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
\code{DefaultAssay}: The name of the default assay

\code{defaultAssay<-}: An object with the default assay updated
}
\description{
Get and set the default assay
}
