% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dxPlot}
\alias{dxPlot}
\title{Plot near-losslessness parameters.}
\usage{
dxPlot(object, sparsity = 0.001, ...)
}
\arguments{
\item{object}{A Qboneobject}

\item{...}{Arguments passed to other methods}
}
\description{
Plots the near-losslessness parameters minimum concordance (\eqn{\rho^{o}})
and average (\eqn{\bar{\rho}}) against the number of basis coefficients
function of \eqn{K_{c}} in the reduced set.
}
