% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenmapmm}
\alias{eigenmapmm}
\title{Faster Matrix Multiplication (Eigen::Map)}
\usage{
eigenmapmm(A, B)
}
\arguments{
\item{A, B}{Matrices.}
}
\description{
Faster matrix multiplication using C++ Eigen::Map.
A * B
}
