% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenmapmmt}
\alias{eigenmapmmt}
\title{Faster Matrix Multiplication (Eigen::Map)}
\usage{
eigenmapmmt(A, B)
}
\arguments{
\item{A, B}{Matrices.}
}
\description{
Faster matrix multiplication using C++ Eigen::Map.
A * t(B)
}
