% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenmapmt}
\alias{eigenmapmt}
\title{Faster Matrix Transpose (Eigen::Map)}
\usage{
eigenmapmt(A)
}
\arguments{
\item{A}{Matrices.}
}
\description{
Faster matrix multiplication using C++ Eigen::Map
t(A)
}
