% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenmapmtm}
\alias{eigenmapmtm}
\title{Faster Matrix Multiplication (Eigen::Map)}
\usage{
eigenmapmtm(A, B)
}
\arguments{
\item{A, B}{Matrices.}
}
\description{
Faster matrix multiplication using C++ Eigen::Map.
t(A) * B
}
