% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenmm}
\alias{eigenmm}
\alias{eigenmt}
\alias{eigenmapmt}
\alias{eigenmapmm}
\alias{eigenmapmmt}
\alias{eigenmapmtm}
\title{Faster Matrix Multiplication}
\usage{
eigenmm(A, B)

eigenmt(A)

eigenmapmt(A)

eigenmapmm(A, B)

eigenmapmmt(A, B)

eigenmapmtm(A, B)
}
\arguments{
\item{A}{Matrices.}

\item{A, B}{Matrices.}
}
\description{
Faster matrix multiplication using C++ Eigen.
A * B

Faster matrix multiplication using C++ Eigen.
t(A)

Faster matrix multiplication using C++ Eigen::Map
t(A)

Faster matrix multiplication using C++ Eigen::Map.
A * B

Faster matrix multiplication using C++ Eigen::Map.
A * t(B)

Faster matrix multiplication using C++ Eigen::Map.
t(A) * B
}
