% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{idents}
\alias{idents}
\alias{idents<-}
\alias{idents.Qbone}
\alias{idents<-.Qbone}
\alias{levels.Qbone}
\alias{levels<-.Qbone}
\alias{droplevels.Qbone}
\title{Get, set, and manipulate an object's identity classes}
\usage{
idents(object, ...)

idents(object, ...) <- value

\method{idents}{Qbone}(object, ...)

\method{idents}{Qbone}(object, samples = NULL, drop = FALSE, ...) <- value

\method{levels}{Qbone}(x)

\method{levels}{Qbone}(x) <- value

\method{droplevels}{Qbone}(x, ...)
}
\arguments{
\item{object}{An Qbone object}

\item{...}{Arguments passed to other methods}

\item{value}{The name of the identities to pull from object metadata or the
identities themselves}

\item{samples}{Set cell identities for specific samples}

\item{drop}{Drop unused levels}
}
\value{
\code{idents<-}: \code{object} with the cell identities changed
}
\description{
Get, set, and manipulate an object's identity classes
}
\examples{
\dontrun{
# Get sample identity classes
idents(qbone1)
}

\dontrun{
# Set sample identity classes
# Can be used to set identities for specific samples to a new level
idents(qbone1, samples = 3:4) <- 'c'
idents(qbone1)

# Can also set idents from a value in object metadata
colnames(qbone1[[]])
idents(qbone1) <- colnames(qbone1[[]])[3]
# or
idents(qbone1) <- 'group2'
idents(qbone1)
}

\dontrun{
# Get the levels of identity classes of a Qbone object
levels(x = qbone1)
}

\dontrun{
# Reorder identity classes
levels(x = qbone1)
levels(x = qbone1) <- c('A', 'B')
levels(x = qbone1)
}

}
