% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunction.R
\name{lassolist2}
\alias{lassolist2}
\title{Run lasso list to build basis function}
\usage{
lassolist2(
  object,
  verbose = F,
  new.assay.name = "Lasso.list",
  alpha = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  beta = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  assay.seed = .Random.seed,
  parallel = T,
  ...
)
}
\arguments{
\item{object}{A Qboneobject}

\item{verbose}{Print a progress bar}

\item{new.assay.name}{new assay name assigned to the lassolist data}

\item{alpha}{vector containing sequence of beta parameter for internal function \code{generateBetaCDF()}}

\item{beta}{vector containing sequence of beta parameter for internal function \code{generateBetaCDF()}}

\item{assay.seed}{assay information to add into the QboneData object scale.data. The default of \code{lassolist()} will save the random seed for the run. \code{assay.seed = object@assays[["Lasso.list"]]@scale.data[["lassolist"]]} to run \code{lassolist()} for the same results.}

\item{parallel}{If TRUE, use parallel foreach to fit each fold. Must register parallel before hand}

\item{...}{Arguments passed to other methods}
}
\description{
Run lasso list to build basis function
}
