% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{nllplot}
\alias{nllplot}
\title{Plot near-losslessness parameters.}
\usage{
nllplot()
}
\arguments{
\item{object}{A Qboneobject}
}
\description{
Plots the near-losslessness parameters minimum concordance (\eqn{\rho^{o}})
and average (\eqn{\bar{\rho}}) against the number of basis coefficients
function of \eqn{K_{c}} in the reduced set.
}
