% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{nllplotrev}
\alias{nllplotrev}
\title{Plot number of basis coefficients.}
\usage{
nllplotrev(object, ...)
}
\arguments{
\item{object}{A Qboneobject#'}

\item{...}{Arguments passed to other methods}
}
\description{
Reversed nllplot useful when plotting with other basis like principal
components.
minimum concordance (\eqn{\rho^{o}}) and average (\eqn{\bar{\rho}})
for quantlets basis, varying with the number of basis coefficients.
}
