% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneObject.R
\name{assays}
\alias{assays}
\title{Query Specific Object Types}
\usage{
assays(object, slot = NULL)
}
\arguments{
\item{object}{A \code{\link{Qbone}} object}

\item{slot}{Name of component object to return}
}
\value{
If \code{slot} is \code{NULL}, the names of all component objects
in this \code{Qbone} object. Otherwise, the specific object specified
}
\description{
Adopted from Seurateobject package
List the names of \code{\link{assay}}
}
\concept{data-access}
