% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pdPlot}
\alias{pdPlot}
\title{Plot predicted densities}
\usage{
pdPlot(
  object,
  plot.col,
  group.names,
  mean.diff = F,
  var.diff = F,
  skewed.diff = F,
  kurtosis.diff = F,
  ...
)
}
\arguments{
\item{object}{A Qboneobject}

\item{plot.col}{Columns to plot from \code{qfrModel()} results \code{object@assays[["Q.F.Regression"]]@scale.data[["mcmc_infer"]][["den_G"]] }. Corresponding to \code{X1} (new covariates) agrement in \code{qfrModel()}.}

\item{group.names}{Group name for plot legend for argument \cond{plot.col}.}

\item{mean.diff}{T or F to add mean difference testing for two consecutive subjects posterior probability scores}

\item{var.diff}{T or F to add variance difference testing for two consecutive subjects posterior probability scores}

\item{skewed.diff}{T or F to add skewness difference testing for two consecutive subjects posterior probability scores}

\item{kurtosis.diff}{T or F to add kurtosis difference testing for two consecutive subjects posterior probability scores}

\item{...}{Arguments passed to other methods}
}
\description{
Plot predicted densities
}
