% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunction.R
\name{preQuantlets}
\alias{preQuantlets}
\title{Get the pre-quantlets basis functions}
\usage{
preQuantlets(
  object,
  new.assay.name = "Pre.Quantiles",
  data.assay = defaultAssay(object),
  p = signif(seq(0.001, 0.999, length = 1024), 4),
  alpha = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  beta = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  ...
)
}
\arguments{
\item{object}{A Qboneobject}

\item{new.assay.name}{New assay name assigned to the quantlets data}

\item{data.assay}{It is the name of the assay whose data will be used
to compute the lasso list. Default is the data from the
defaultAssay(object).}

\item{p}{Vector of length P  in (0,1)    Probability grids.}

\item{alpha}{Vector containing sequence of beta parameter for internal
function \code{generateBetaCDF()}}

\item{beta}{Vector containing sequence of beta parameter for internal
function \code{generateBetaCDF()}}

\item{...}{Arguments passed to other methods}
}
\description{
Take union set and find near-lossless subset as confirmed by
cross-validated concordance correlation coefficient (CVCCC)
}
