% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneObject.R
\docType{class}
\name{Qbone-class}
\alias{Qbone-class}
\alias{Qbone}
\title{The Qbone Class}
\description{
The Qbone object is a representation of data analysis using Quantile
Functional Regression using Quantlets (doi: 10.1080/01621459.2019.1609969)
for R.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays data for this project.
structure like assay name
├── sample list
│        ├── sample 1
│        ├── sample 2
...      ...
│        └── sample x
└── assay information}

\item{\code{meta.data}}{Meta information regarding each sample.}

\item{\code{active.assay}}{Name of the active, or default, assay; settable using
\code{\link{defaultAssay}}}

\item{\code{active.ident}}{The active identity for this Qbone object;
settable using \code{\link{idents}}}

\item{\code{graphs}}{list of \code{\link{Graph}} objects                                 || double check}

\item{\code{images}}{A list of image objects                                             || double check}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of miscellaneous information                                   || double check}

\item{\code{version}}{Version of Qbone this object was built under}

\item{\code{commands}}{A list of logged commands run on this \code{Qbone} object        || double check}

\item{\code{tools}}{A list of miscellaneous data generated by other tools, should be     || double check
filled by developers only using \code{\link{Tool}<-}}
}}

