% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneObject.R
\name{Qbone-methods}
\alias{Qbone-methods}
\alias{[[.Qbone}
\alias{[[<-,Qbone-method}
\title{Qbone Methods}
\usage{
\method{[[}{Qbone}(x, i, ..., drop = FALSE)

\S4method{[[}{Qbone}(x, i, j, ...) <- value
}
\arguments{
\item{x, object}{A \code{\link{Qbone}} object}

\item{i}{Depends on the method
\describe{
\item{\code{[}, \code{subset}}{Feature names or indices}
\item{\code{$}, \code{$<-}}{Name of a single metadata column}
\item{\code{[[}, \code{[[<-}}{
Name of one or more metadata columns or an associated object; associated
objects include \code{\link{Assay}}, \code{\link{DimReduc}},                  || double check
\code{\link{Graph}}, \code{\link{SeuratCommand}}, or
\code{\link{SpatialImage}} objects
}
}}

\item{...}{Arguments passed to other methods}

\item{drop}{See \code{\link[base]{drop}}}

\item{j, samples}{Sample names or indices}

\item{value}{Additional metadata or associated objects to add; \strong{note}:
can pass \code{NULL} to remove metadata or an associated object}

\item{n}{The number of rows of metadata to return}
}
\value{
\code{[[}: If \code{i} is missing, the metadata data frame; if
\code{i} is a vector of metadata names, a data frame with the requested
metadata, otherwise, the requested associated object

\code{[[<-}: \code{x} with the metadata or associated objects added
as \code{i}; if \code{value} is \code{NULL}, removes metadata or associated
object \code{i} from object \code{x}
}
\description{
Methods for \code{\link{Qbone}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{[[.Qbone}: Metadata and associated object accessor

\item \code{[[<-,Qbone-method}: Metadata and associated object accessor
}}

