% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{readQbone}
\alias{readQbone}
\title{Read Qbone from files}
\usage{
readQbone(data.dir, groupbyfolder = F, data.column = 1)
}
\arguments{
\item{data.dir}{the folder directory where the raw csv file saved}

\item{groupbyfolder}{Use parent folder names as sample group information if set to T, default is F}

\item{data.column}{column number of sample data in the raw csv file}
}
\description{
Read Qbone from files
}
