% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{samples}
\alias{samples}
\alias{samples,Qbone-method}
\title{Get samples present in an object}
\usage{
samples(x)

\S4method{samples}{Qbone}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of sample names
}
\description{
Get samples present in an object
}
