% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set-if-null}
\alias{set-if-null}
\alias{\%||\%}
\title{Set a default value depending on if an object is \code{NULL}
From rlang package.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{An object to test}

\item{y}{A default value}
}
\value{
For \code{\%||\%}: \code{y} if \code{x} is \code{NULL} otherwise
\code{x}
}
\description{
Set a default value depending on if an object is \code{NULL}
From rlang package.
}
\examples{
1 \%||\% 2
NULL \%||\% 2

}
