% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{thinData}
\alias{thinData}
\title{Thin the data
If the data set is too big and the calculation is too heavy for your computer you may use this function to thin the data to make the calculation faster.
Here we use thin the data proportionally on the sorted data. The new data set should reserve the same quantile feature as it was sequentially subsetted from ordered data set.}
\usage{
thinData(object, new.assay.name = "Thin", prop = NULL)
}
\arguments{
\item{object}{An Qboneobject.}

\item{new.assay.name}{new assay name assigned to the thined data}

\item{prop}{proportion to keep from the original data. By default the scale.data of \code{thinData()} will save the prop of thinning}
}
\description{
Thin the data
If the data set is too big and the calculation is too heavy for your computer you may use this function to thin the data to make the calculation faster.
Here we use thin the data proportionally on the sorted data. The new data set should reserve the same quantile feature as it was sequentially subsetted from ordered data set.
}
\concept{Data thin}
