% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneData.R
\name{createQboneData}
\alias{createQboneData}
\title{create an QboneData object}
\usage{
createQboneData(
  data,
  meta.assays = NULL,
  sampleid.assays = 1,
  assay.name = "Bone",
  assay.orig = NULL,
  sort = F,
  ...
)
}
\arguments{
\item{data}{Sample data}

\item{meta.assays}{metadata for the assay with sample names}

\item{sampleid.assays}{column number of sample name in meta.assay}

\item{assay.name}{assay name for the data.}

\item{assay.orig}{Original assay that this assay is based off of. Used to
track assay provenance}

\item{sort}{sort the data before put into data slot if T, default is F.}
}
\value{
A \code{\link{QboneData}} object
}
\description{
create an QboneData object
}
\examples{
\dontrun{
n = 10000
list1 = list(rnorm(n, mean = 0, sd = 1),
             rnorm(n, mean = 0, sd = 1),
            rnorm(n, mean = 0.5, sd = 1),
            rnorm(n, mean = 0.5, sd = 1))
meta.data = data.frame(name = c("a_1", "a_2", "b_1", "b_2"), group = c("a", "a", "b", "b"))
rownames(meta.data) = meta.data[,1]
qbonedata = createQboneData(list1, meta.data, sampleid = 1)
}

}
