% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneData.R
\name{createQboneData}
\alias{createQboneData}
\title{create an QboneData object}
\usage{
createQboneData(
  data,
  meta.assays = NULL,
  sampleid = 1,
  assay.name = "Bone",
  assay.orig = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{meta.assays}{Include cells where at least this many features are
detected.}

\item{sampleid}{column number of sample name in meta.assay}

\item{assay.name}{assay name for the data.}

\item{assay.orig}{Original assay that this assay is based off of. Used to
track assay provenance}

\item{...}{|| double check}
}
\value{
A \code{\link{QboneData}} object
}
\description{
create an QboneData object
}
