% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{createQboneObject}
\alias{createQboneObject}
\alias{createQboneObject.default}
\alias{createQboneObject.QboneData}
\title{Create a \code{Qbone} object}
\usage{
createQboneObject(
  data,
  project = "createQboneObject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  sampleid = 1,
  ...
)

\method{createQboneObject}{default}(
  data = data,
  project = "QboneProject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)

\method{createQboneObject}{QboneData}(
  data = data,
  project = "QboneProject",
  assay = "Bone",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{project}{\link{project} name for the \code{Qbone} object}

\item{assay}{Name of the initial assay}

\item{names.field}{|| double check}

\item{names.delim}{|| double check}

\item{meta.data}{Include cells where at least this many features are
detected.}

\item{sampleid}{column number of sample name in meta.assay}

\item{...}{|| double check}

\item{row.names}{When \code{counts} is a \code{data.frame} or
\code{data.frame}-derived object: an optional vector of feature names to be
used}
}
\description{
Create a \code{Qbone} object
}
