% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneObject.R
\name{idents}
\alias{idents}
\alias{idents<-}
\alias{idents.Qbone}
\alias{idents<-.Qbone}
\title{Get, set, and manipulate an object's identity classes}
\usage{
idents(object, ...)

idents(object, ...) <- value

\method{idents}{Qbone}(object, ...)

\method{idents}{Qbone}(object, samples = NULL, drop = FALSE, ...) <- value
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{value}{The name of the identities to pull from object metadata or the
identities themselves}

\item{samples}{Set cell identities for specific samples}

\item{drop}{Drop unused levels}

\item{x, object}{An object}
}
\value{
\code{idents<-}: \code{object} with the cell identities changed
}
\description{
Get, set, and manipulate an object's identity classes
}
