% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunction.R
\name{lassolist}
\alias{lassolist}
\title{Run lasso list to build basis function}
\usage{
lassolist(
  object,
  verbose = TRUE,
  new.assay.name = "Lasso",
  a1 = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  a2 = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  assay.seed = .Random.seed,
  ...
)
}
\arguments{
\item{object}{A Qboneobject}

\item{verbose}{Print a progress bar}

\item{new.assay.name}{new assay name assigned to the lassolist data}

\item{a1}{vector containing sequence of beta parameter for internal function \code{\link{generateBetaCDF}}}

\item{a2}{vector containing sequence of beta parameter for internal function \code{\link{generateBetaCDF}}}

\item{assay.seed}{assay information to add into the QboneData object scale.data. The default of \code{lassolist()} will save the random seed for the run. \code{.Random.seed <- boject@assays[["Lasso"]]@scale.data[["lassolist"]] } to reset the seed for the same results.}
}
\description{
Run lasso list to build basis function
}
