% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunction.R
\name{lassolist}
\alias{lassolist}
\title{Run lasso list to build basis function}
\usage{
lassolist(
  object,
  verbose = TRUE,
  data.assay = defaultAssay(object),
  new.assay.name = "Lasso.list",
  alpha = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  beta = c(seq(0.1, 1, by = 0.1), seq(2, 100, by = 1)),
  assay.seed = .Random.seed,
  parallel = T,
  ...
)
}
\arguments{
\item{object}{A Qboneobject}

\item{verbose}{Print a progress bar}

\item{data.assay}{It is the name of the assay whose data will be used to compute the lasso list. Default is the data from the defaultAssay(object).}

\item{new.assay.name}{New assay name assigned to the lassolist data}

\item{alpha}{Vector containing sequence of beta parameter for internal function \code{generateBetaCDF()}}

\item{beta}{Vector containing sequence of beta parameter for internal function \code{generateBetaCDF()}}

\item{assay.seed}{assay information to add into the QboneData object scale.data. The default of \code{lassolist()} will save the random seed for the run. Use \code{.Random.seed <-  object@assays[["Lasso.list"]]@scale.data[["lassolist"]]} before run \code{lassolist()} for the same results.}

\item{parallel}{If TRUE, use parallel foreach to fit each fold in \code{cv.glmnet()}. Default use \code{registerDoMC()} to register parallel.}

\item{...}{Arguments passed to other methods}
}
\description{
Run lasso list to build basis function
}
