% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qbasisplot}
\alias{qbasisplot}
\title{Plot first n quantlet basis functions}
\usage{
qbasisplot()
}
\arguments{
\item{object}{A Qboneobject}

\item{n}{Number of first n basis functions to plot, default = 16.}
}
\description{
Plot first n quantlet basis functions
}
