% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneObject.R
\docType{class}
\name{Qbone-class}
\alias{Qbone-class}
\alias{Qbone}
\title{The Qbone Class}
\description{
The Qbone object is a representation of data analysis using Quantile
Functional Regression using Quantlets (doi: 10.1080/01621459.2019.1609969)
for R.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of QboneData Object for this project.}

\item{\code{meta.data}}{Meta information regarding each sample.}

\item{\code{active.assay}}{Name of the active, or default, assay; settable using
\code{\link{defaultAssay}}}

\item{\code{active.ident}}{The active identity for this Qbone object;
settable using \code{\link{idents}}}

\item{\code{project.name}}{Name of the project}

\item{\code{version}}{Version of Qbone this object was built under}
}}

