% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Qbone-package}
\alias{Qbone-package}
\alias{_PACKAGE}
\title{Qbone: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Simin Pan \email{simonpan@tamu.edu} (\href{https://orcid.org/0000-0003-3472-8940}{ORCID})

}
