% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Qbone-package}
\alias{Qbone-package}
\alias{_PACKAGE}
\title{Qbone: Tools for Quantile Functional Regression with Quantlets}
\description{
The `Qbone` package is a set of tools that enables efficient estimation of quantlets used in quantile functional regression to analyze distributional data. This package can be used to analyze distributional data in broader applications such as climate change analysis, image recognition analysis, and personalized health data analysis, etc. At current stage, it is just for independent functional linear regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/siminpan/Qbone}
  \item Report bugs at \url{https://github.com/siminpan/Qbone/issues}
}

}
\author{
\strong{Maintainer}: Simin Pan \email{simonpan@tamu.edu} (\href{https://orcid.org/0000-0003-3472-8940}{ORCID})

Authors:
\itemize{
  \item Linying Zhang \email{zly.rebecca@tamu.edu} (\href{https://orcid.org/0000-0001-9658-9427}{ORCID})
}

Other contributors:
\itemize{
  \item Alan Dabney \email{adabney@stat.tamu} (\href{https://orcid.org/0000-0002-1240-4171}{ORCID}) [thesis advisor]
  \item Carl Gregory \email{cgregory@tamu.edu} (\href{https://orcid.org/0000-0001-6444-5541}{ORCID}) [thesis advisor]
}

}
