% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneAssay.R
\docType{class}
\name{QboneAssay-class}
\alias{QboneAssay-class}
\alias{QboneAssay}
\title{The QboneAssay Class
The QboneAssay object is the basic unit of Qbone.}
\description{
The QboneAssay Class
The QboneAssay object is the basic unit of Qbone.
}
\seealso{
\code{\link{QboneAssay-methods}}
}
