% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneData.R
\docType{class}
\name{QboneData-class}
\alias{QboneData-class}
\alias{QboneData}
\title{The QboneData Class
The QboneData object is the basic unit of Qbone.}
\description{
The QboneData Class
The QboneData object is the basic unit of Qbone.
}
\section{Slots}{

\describe{
\item{\code{data}}{Sample data}

\item{\code{scale.data}}{Parameter of data processing}

\item{\code{key}}{Key for the Assay                                                || double check}

\item{\code{assay.name}}{name of assay}

\item{\code{assay.orig}}{Original assay that this assay is based off of. Used to
track assay provenance}

\item{\code{meta.assays}}{Metadata for the data processing}

\item{\code{misc}}{Utility slot for storing additional data associated with the assay   || double check}
}}

\seealso{
\code{\link{QboneData-methods}}
}
