% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QboneData.R
\name{QboneData-methods}
\alias{QboneData-methods}
\title{QboneData Methods
\code{QboneData} Methods}
\arguments{
\item{x, object}{An \code{\link{QboneData}} object}

\item{i, features}{For \item{\code{[[}}: metadata names; for all other methods,
feature names or indices}

\item{j, samples}{sample names or indices}

\item{...}{Arguments passed to other methods}
}
\description{
Methods for \code{\link{QboneData}} objects for generics defined in
other packages
}
\concept{QboneData}
