% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/QboneData.R, R/QboneObject.R
\name{samples}
\alias{samples}
\alias{samples.QboneData}
\alias{samples.Qbone}
\title{Get samples present in an object}
\usage{
samples(x)

\method{samples}{QboneData}(x)

\method{samples}{Qbone}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of sample names
}
\description{
Get samples present in an object
}
\examples{
samples(x = q1)


}
